It took me about 15 tweaks to figure this out
First problem:  Figuring out that Python sees all input as a string and changes it back to a string after it performs a math function with it.  It does have no problems having a variable that changes value.
Second problem: My math background complicates how I understand how a computer thinks...I over complicated the even and odd where the programmer I  used for reference just had it evaluate the number for divisive by 2 to be an interger or not...and made it a yes /no question...if it has a decimal it is odd and not an interger so it comes back zero.
Third Problem: Grasping how to position break and where to put the number count in relation to while statements.